+++ AlphaWallet/Tokens/Types/TokenObject.swift
@@ -122,31 +122,6 @@ class TokenObject: Object {
     var server: RPCServer {
         return .init(chainID: chainId)
     }
-
-    var iconImage: UIImage? {
-        switch type {
-        case .nativeCryptocurrency:
-            return server.iconImage
-        case .erc20, .erc875, .erc721, .erc721ForTickets:
-            return nil
-        }
-    }
-
-    var symbolBackgroundColor: UIColor {
-        if contractAddress.sameContract(as: Constants.nativeCryptoAddressInDatabase) {
-            return server.blockChainNameColor
-        } else {
-            let colors = [R.color.radical()!, R.color.cerulean()!, R.color.emerald()!, R.color.indigo()!, R.color.azure()!, R.color.pumpkin()!]
-            let index: Int
-            //We just need a random number from the contract. The LSBs are more random than the MSBs
-            if let i = Int(contractAddress.eip55String.substring(from: 37), radix: 16) {
-                index = i % colors.count
-            } else {
-                index = 0
-            }
-            return colors[index]
-        }
-    }
 }
 
 func isNonZeroBalance(_ balance: String) -> Bool {
